Adapt to changes in Kdevelop
Specifically changes 08393593 and 32a55f871. Basically, use a 64 bit value on 64 bit platforms

+++ duchain/types/hintedtype.cpp
@@ -97,7 +97,7 @@ bool HintedType::equals(const AbstractType* rhs) const
     return true;
 }
 
-uint HintedType::hash() const
+size_t HintedType::hash() const
 {
     return AbstractType::hash() + 1 + ( type() ? type()->hash() : 0 ) + d_func()->m_createdByContext.index()
                                 + d_func()->m_modificationRevision.modificationTime % 17 + (d_func()->m_modificationRevision.revision * 19) % 13;
